git-svn-id: file:///home/yoann/SVN/trunk/libprelude@2691 09c5ec92-17d4-0310-903a-819935f44dba

+++ src/daemonize.c
@@ -128,7 +128,7 @@ static int lockfile_write_pid(int fd, pid_t pid)
                 return -1;
         }
         
-        snprintf(buf, sizeof(buf), "%d\n", pid);
+        snprintf(buf, sizeof(buf), "%d\n", (int) pid);
         
         ret = write(fd, buf, strlen(buf));
         if ( ret < 0 ) {
